use aliases; move code to @db and @serverFunctions files remove duplicated code
