+++ src/db/actuality.ts
@@ -1,16 +1,9 @@
 import { Prisma } from '@prisma/client'
 import { prismaClient } from './client'
 
-export const getAllActualities = () =>
-  prismaClient.actuality.findMany({
-    orderBy: { createdAt: 'desc' },
-  })
+export const getAllActualities = () => prismaClient.actuality.findMany({ orderBy: { createdAt: 'desc' } })
 
-export const getMainActualities = () =>
-  prismaClient.actuality.findMany({
-    orderBy: { createdAt: 'desc' },
-    take: 3,
-  })
+export const getMainActualities = () => prismaClient.actuality.findMany({ orderBy: { createdAt: 'desc' }, take: 3 })
 
 export const createActualities = async (data: Prisma.ActualityCreateInput[]) =>
   prismaClient.actuality.createMany({ data })
