+++ src/scripts/oldBC/transition/emissionFactors.ts
@@ -1,8 +1,8 @@
+import { OrganizationVersionWithOrganization } from '@/db/organization'
+import { unitsMatrix } from '@/services/importEmissionFactor/historyUnits'
+import { getEmissionQuality } from '@/services/importEmissionFactor/import'
 import { EmissionFactorPartType, EmissionFactorStatus, Import, Prisma } from '@prisma/client'
 import { v4 } from 'uuid'
-import { OrganizationVersionWithOrganization } from '../../../db/organization'
-import { unitsMatrix } from '../../../services/importEmissionFactor/historyUnits'
-import { getEmissionQuality } from '../../../services/importEmissionFactor/import'
 import { EmissionFactorsWorkSheet } from './oldBCWorkSheetsReader'
 
 const getStringValue = (value: string | number) => {
