+++ src/scripts/oldBC/transition/oldBC.ts
@@ -1,11 +1,11 @@
 import { getAccountByEmailAndOrganizationVersionId } from '@/db/account'
+import { prismaClient } from '@/db/client'
 import {
   getOrganizationVersionById,
   getOrganizationWithSitesById,
   OrganizationVersionWithOrganization,
 } from '@/db/organization'
 import { Environment } from '@prisma/client'
-import { prismaClient } from '../../../db/client'
 import { uploadEmissionFactors } from './emissionFactors'
 import { OldNewPostAndSubPostsMapping } from './newPostAndSubPosts'
 import { OldBCWorkSheetsReader } from './oldBCWorkSheetsReader'
